Extract IntervalTrigger from trigger module and put it under triggers namespace
