+++ chainer/training/triggers/minmax_value_trigger.py
@@ -1,5 +1,5 @@
 from chainer import reporter
-from chainer.training import trigger as trigger_module
+from chainer.training.triggers import interval
 
 
 class BestValueTrigger(object):
@@ -20,7 +20,7 @@ class BestValueTrigger(object):
     def __init__(self, key, compare, trigger=(1, 'epoch')):
         self._key = key
         self._best_value = None
-        self._interval_trigger = trigger_module.get_trigger(trigger)
+        self._interval_trigger = interval.get_trigger(trigger)
         self._init_summary()
         self._compare = compare
